pyproject.toml
[project]
name = "posture-system-py39"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.9.25"
dependencies = [
    "adafruit-blinka>=9.0.4",
    "adafruit-circuitpython-framebuf>=1.6.11",
    "adafruit-circuitpython-ssd1306>=2.12.23",
    "adafruit-ssd1306>=1.6.2",
    "board>=1.0",
    "flask>=3.1.3",
    "gtts>=2.5.4",
    "hx711>=1.1.2.3",
    "mediapipe-rpi3>=0.8.8",
    "opencv-python>=4.13.0.92",
    "pygame>=2.6.1",
    "rpi-gpio>=0.7.1",
]
